+++ CHANGELOG.md
@@ -7,11 +7,13 @@
 * Increase `Suite`, `Run` test coverage
 * `TestrailRun#is_completed` accessor to check for closed runs
 * `TestrailProject#suites` to get list of Suites
+* `TestrailProject#runs` to get list of Runs
 
 ### Changes
 
 * Do not send `codecov` info on non-CI runs
 * Deprecate `TestrailProject#get_suites`
+* Deprecate `TestrailProject#get_runs`
 
 ## 0.1.0 (2020-11-27)